Create a function for inserting each individual tag
This patch expands the API for the mode by creating functions like
bbcode/insert-tag-b, bbcode/insert-tag-code, and so on for all of the
tags bound to keys via bbcode/make-key-binding.  The makes it possible
to insert specific tags easily in other Elisp code.  And it makes the
command ‘C-h m’ more useful by showing every function that every key
uses in the mode.