Call use-region-p instead of testing the value of mark-active
The Elisp manual says this is a better way to determine whether or not
we should apply the effects of a command to a selected region.
However, this patch still does not fix the bug by which no tag
insertion commands will work until the user has created a mark.