Fix two bugs in bbcode/insert-tag
`kill-region` and `yank` modify the kill ring (clipboard) and should
therefore only be used by the user directly. Elisp code is supposed to
use similar functions such as `buffer-substring` and `delete-region`
which leave the kill ring intact. We can also avoid the `use-region-p`
check in this function since `buffer-substring`, `delete-region` and
their ilk use buffer offset arguments (start and end) instead of
looking up the region themselves.

`search-backward` is not an appropriate function to use when we are
looking for a known string that we just inserted into a known position
in the buffer. Instead, we should simply use `goto-char` and calculate
the desired buffer position from known string lengths. In this case,
if the user's region contained a `]` character, then `search-backward`
would mistakenly think that the `]` in the region was the end of the
opening tag we inserted.